# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Conecta API",
    "version": "0.1.0",
    "description": "Hackathon Conexão Ancestral | Petronect + KODIE Academy | Equipe 05. Base exclusivamente simulada. Admin exige Bearer no modo local; a alternativa sem autenticação se aplica somente a GET em DEMO_READ_ONLY=true."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Desenvolvimento local"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              }
            },
            "required": ["code", "message", "requestId"]
          }
        },
        "required": ["error"]
      },
      "Profile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "enum": ["demo-01", "demo-02", "demo-03", "demo-04", "demo-05", "demo-06"]
          },
          "label": {
            "type": "string"
          },
          "segment": {
            "type": "string",
            "enum": ["energia", "tecnologia", "servicos"]
          }
        },
        "required": ["id", "label", "segment"]
      },
      "SessionInput": {
        "type": "object",
        "properties": {
          "profileId": {
            "$ref": "#/components/schemas/ProfileId"
          },
          "analyticsConsent": {
            "const": true
          }
        },
        "required": ["profileId", "analyticsConsent"],
        "additionalProperties": false
      },
      "ProfileId": {
        "type": "string",
        "enum": ["demo-01", "demo-02", "demo-03", "demo-04", "demo-05", "demo-06"]
      },
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string"
          },
          "expiresIn": {
            "const": 86400
          },
          "simulated": {
            "const": true
          }
        },
        "required": ["id", "token", "expiresIn", "simulated"]
      },
      "Preference": {
        "type": "object",
        "properties": {
          "analyticsConsent": {
            "type": "boolean"
          }
        },
        "required": ["analyticsConsent"],
        "additionalProperties": false
      },
      "EventInput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID v4; idempotente globalmente"
          },
          "type": {
            "type": "string",
            "enum": ["page_view", "click", "preference", "journey_completed"]
          },
          "page": {
            "type": "string",
            "enum": ["home", "empresa", "oportunidades", "ajuda", "preferencias"]
          },
          "target": {
            "type": "string",
            "enum": ["page", "explorar", "ajuda", "energia", "tecnologia", "servicos", "concluir"]
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time",
            "description": "Últimas 24h; tolerância de 60 segundos no futuro"
          }
        },
        "required": ["id", "type", "page", "target", "occurredAt"],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "adminBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "ADMIN_TOKEN local de pelo menos 32 caracteres"
      },
      "sessionBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token retornado por POST /sessions; válido por 24h"
      }
    }
  },
  "paths": {
...
    "/api/v1/catalog": {
      "get": {
        "summary": "Catálogo fictício permitido",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "simulated": {
                      "type": "boolean"
                    },
                    "profiles": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Profile"
                      }
                    },
                    "interests": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "readOnly": {
                      "type": "boolean"
                    }
                  },
                  "required": ["simulated", "profiles", "interests", "readOnly"]
                }
              }
            }
          },
          "400": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions": {
      "post": {
        "summary": "Cria sessão fictícia; exige modo de escrita",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{id}/context": {
      "get": {
        "summary": "Contexto mínimo da própria sessão",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profileId": {
                      "type": "string"
                    },
                    "analyticsConsent": {
                      "type": "boolean"
                    },
                    "simulated": {
                      "type": "boolean"
                    },
                    "nextStep": {
                      "type": "string"
                    }
                  },
                  "required": ["profileId", "analyticsConsent", "simulated", "nextStep"]
                }
              }
            }
          },
          "400": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/sessions/{id}/preferences": {
      "patch": {
        "summary": "Altera coleta; retirada apaga eventos da sessão",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "analyticsConsent": {
                      "type": "boolean"
                    },
                    "eventsRemoved": {
                      "type": "boolean"
                    }
                  },
                  "required": ["analyticsConsent", "eventsRemoved"]
                }
              }
            }
          },
          "400": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preference"
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{id}/events": {
      "post": {
        "summary": "Registra evento único ou reconhece repetição idêntica",
        "security": [
          {
            "sessionBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Repetição idêntica",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accepted": {
                      "type": "boolean"
                    },
                    "duplicate": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": ["accepted", "duplicate", "id"]
                }
              }
            }
          },
          "201": {
            "description": "Novo evento",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accepted": {
                      "type": "boolean"
                    },
                    "duplicate": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": ["accepted", "duplicate", "id"]
                }
              }
            }
          },
          "400": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/summary": {
      "get": {
        "summary": "Métricas do recorte completo",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Summary"
                }
              }
            }
          },
          "400": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro conforme docs/API.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "adminBearer": []
          },
          {}
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Início inclusivo UTC; padrão 30 dias antes de hoje"
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Fim inclusivo UTC; padrão hoje; intervalo máximo 366 dias"
          },
          {
            "name": "segment",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["energia", "tecnologia", "servicos"]
            }
          }
        ]
      }
    },
...
  }
}